mailman: split patchset handling into separate file
